Catch some more missing long conversions.

+++ identity/src/main/java/com/android/identity/SessionEncryptionReader.java
@@ -126,7 +126,7 @@ final class SessionEncryptionReader {
         if (!(cipherSuiteDataItem instanceof Number)) {
             throw new IllegalArgumentException("Cipher suite not a Number");
         }
-        int cipherSuite = ((Number) cipherSuiteDataItem).getValue().intValue();
+        long cipherSuite = ((Number) cipherSuiteDataItem).getValue().longValue();
         if (cipherSuite != 1) {
             throw new IllegalArgumentException("Expected cipher suite 1, got " + cipherSuite);
         }
